requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djoser
drf-yasg
django-cors-headers
